+++ noodles-vcf/CHANGELOG.md
@@ -14,13 +14,14 @@
 
   * vcf: Move `lazy::Record` to `Record`.
 
-  * vcf/io/reader: Rename record to record buf.
+  * vcf/io/reader: Rename record to record buf and  lazy record to record.
 
     This changes the following:
 
       * `Reader::read_record` => `Reader::read_record_buf`,
-      * `Reader::records` => `Reader::record_bufs`, and
-      * `Records` => `RecordBufs`.
+      * `Reader::records` => `Reader::record_bufs`,
+      * `Records` => `RecordBufs`, and
+      * `Reader::read_lazy_record` => `Reader::read_record`.
 
 ## 0.51.0 - 2024-03-28